Move state and props into new file
