+++ components/src/BloomreachProductRecommendationsDisplayMain/types/bloomreach.productrecommendations.main.d.ts
@@ -1,0 +1,8 @@
+export interface BloomreachProductRecommendationsDisplayMainProps {
+    cortexProductData: any,
+  }
+
+export interface BloomreachProductRecommendationsDisplayMainState {
+    maxItemsInView: number,
+    productData: any
+}
